src/test/java/net.thumbtack/controller/CategoriesControllerTest.java
package net.thumbtack.controller;

import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.hasSize;
import static org.junit.Assert.assertEquals;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;

import java.util.Arrays;
import java.util.List;
import net.thumbtack.dto.CategoryResponseDto;
import net.thumbtack.model.Category;
import net.thumbtack.service.iface.CategoryService;
import org.junit.Before;
import org.junit.Ignore;
import org.junit.Test;
import org.mockito.ArgumentCaptor;
import org.mockito.Captor;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

public class CategoriesControllerTest {

  private CategoriesController underTest;

  @Mock
  private CategoryService categoryService;

  @Captor
  private ArgumentCaptor<Category> captor;

  @Before
  public void setUpClass() {
    MockitoAnnotations.initMocks(this);
    underTest = new CategoriesController(this.categoryService);
  }

  @Test
  public void testAddCategory() {
    Category category = new Category( 1, "Trausers", 1);
    underTest.addCategory(category);
    verify(categoryService).addCategory(captor.capture());
    Category value = captor.getValue();
    assertEquals(category.getId(), value.getId());
  }


  @Test
  public void testGetAllCategoryAndDeleteCategory() {
    List<Category> categories = Arrays.asList(
        new Category( 1, "Trausers", 1),
        new Category( 2, "Trausers", 1)
    );
    when(categoryService.getAllCategory()).thenReturn(categories);
    List<CategoryResponseDto> categoryList = underTest.getAllCategories().getBody();
    assertThat(categoryList, hasSize(categories.size()));

    underTest.deleteCategory(1);//в постмане всё работает, что сделал неправильно???
    List<CategoryResponseDto> categoryList1 = underTest.getAllCategories().getBody();
    assertThat(categoryList1, hasSize(categories.size()));
  }

  @Ignore
  @Test
  public void testGetCategoryById() {
    Category category = new Category( 1, "Trausers", 1);
    underTest.getCategoryById(1);//в постмане всё работает, что сделал неправильно???
    verify(categoryService).getCategoryById(1);
    Category value = captor.getValue();
    assertEquals(category.getId(), value.getId());
  }



}
